refactor(parser): include annotations contextually
This makes it so annotations can be fetched by the parser when, and only
when, it is expecting an annotation - rather than tying this
admissability to TokenKind (e.g., ParenL can be used in more contexts
than function calls).

Evidently, this is only possible now that we have migrated from
Peekable<Lexer<'a>> to a real TokenStream<'a>, since it allows much more
flexibility.